Added path finding and mouse support
Path finding is implemented on A* algorithm with several heuristics available (manhattan, diagonal
shourtcut and dijkstra).

Mouse support includes drawing a marker over a tile and drawing a movement path for player. Player
can click any tile that he/she can _possibly_ move onto (i.e. all tiles that are not obstacles) and
a correct path will be calculated, but won't be able to move there as long as there is an _obstacle
object_ on that tile.